[Messenger] added a SenderLocator decoupled from ContainerInterface
